tweaks: deleting re imports
