+++ docs/design-plans/2026-04-08-i18n-setup.md
@@ -13,8 +13,8 @@ Full internationalization (i18n) of the Bobadilla Tech marketing site using
 Portuguese (`pt`). All locale-specific URLs are prefixed (`/en/`, `/es/`,
 `/pt/`). Middleware redirects un-prefixed URLs to `/en/` by default.
 
-No functional changes — all features preserved. Pure text extraction and
-routing layer addition.
+No functional changes — all features preserved. Pure text extraction and routing
+layer addition.
 
 ---
 
@@ -66,7 +66,8 @@ safety and make translator diffs readable.
 
 ### Components (UI Chrome)
 
-All components below had hardcoded English strings replaced with `useTranslations()` (client) or `getTranslations()` (async server).
+All components below had hardcoded English strings replaced with
+`useTranslations()` (client) or `getTranslations()` (async server).
 
 | Component                                                     | Namespace            | Key strings                                                        |
 | ------------------------------------------------------------- | -------------------- | ------------------------------------------------------------------ |
@@ -119,15 +120,22 @@ ServiceTechStack, ServiceHero
 
 ### `t.raw()` for arrays
 
-`t()` returns `string` only. Arrays of objects (stats, FAQs, rows, plans, features, benefits) are stored in JSON and retrieved via `t.raw("key") as Type[]`. Icons and prices that can't be serialized to JSON are kept as separate static arrays indexed in parallel.
+`t()` returns `string` only. Arrays of objects (stats, FAQs, rows, plans,
+features, benefits) are stored in JSON and retrieved via
+`t.raw("key") as Type[]`. Icons and prices that can't be serialized to JSON are
+kept as separate static arrays indexed in parallel.
 
 ### Contact form Zod schema
 
-Zod validation messages are set at schema creation time. Moving `contactSchema` inside the component body (re-created on each render) allows `t()` calls to provide locale-aware validation error strings.
+Zod validation messages are set at schema creation time. Moving `contactSchema`
+inside the component body (re-created on each render) allows `t()` calls to
+provide locale-aware validation error strings.
 
 ### Pricing calculator dynamic keys
 
-`PRICING_STEPS` constants retain price/id data. Translated titles, descriptions, and option names live in `messages.PricingPage.steps` and are accessed with template literal keys:
+`PRICING_STEPS` constants retain price/id data. Translated titles, descriptions,
+and option names live in `messages.PricingPage.steps` and are accessed with
+template literal keys:
 
 ```tsx
 t(`steps.${currentStep}.options.${option.id}.name` as Parameters<typeof t>[0]);
@@ -135,25 +143,29 @@ t(`steps.${currentStep}.options.${option.id}.name` as Parameters<typeof t>[0]);
 
 ### Footer nav link translation
 
-`FOOTER_LINKS` is a typed const. A parallel `FOOTER_LINK_KEYS` array maps indices to translation keys, avoiding modifications to the constants file:
+`FOOTER_LINKS` is a typed const. A parallel `FOOTER_LINK_KEYS` array maps
+indices to translation keys, avoiding modifications to the constants file:
 
 ```tsx
 const FOOTER_LINK_KEYS = [
-	"ourServices",
-	"projects",
-	"dedicatedTeam",
-	"openSource",
-	"pricing",
-	"contact",
+  "ourServices",
+  "projects",
+  "dedicatedTeam",
+  "openSource",
+  "pricing",
+  "contact",
 ] as const;
 {
-	FOOTER_LINKS.map((link, i) => t(`links.${FOOTER_LINK_KEYS[i]}`));
+  FOOTER_LINKS.map((link, i) => t(`links.${FOOTER_LINK_KEYS[i]}`));
 }
 ```
 
 ### `t.rich()` for inline markup
 
-`ServiceEstimateCTA` uses `t.rich("body", { gold: chunks => <span className="text-brand-gold">{chunks}</span> })` to render gold-highlighted spans from translated strings without embedding JSX in JSON.
+`ServiceEstimateCTA` uses
+`t.rich("body", { gold: chunks => <span className="text-brand-gold">{chunks}</span> })`
+to render gold-highlighted spans from translated strings without embedding JSX
+in JSON.
 
 ---
 
@@ -161,19 +173,21 @@ const FOOTER_LINK_KEYS = [
 
 ```ts
 matcher: [
-	"/((?!_next/static|_next/image|favicon\\.svg|assets|api|.*\\.[^/]+$).*)",
-	"/",
+  "/((?!_next/static|_next/image|favicon\\.svg|assets|api|.*\\.[^/]+$).*)",
+  "/",
 ];
 ```
 
-API routes (`/api/*`) and static files are explicitly excluded — they never receive a locale prefix.
+API routes (`/api/*`) and static files are explicitly excluded — they never
+receive a locale prefix.
 
 ---
 
 ## Verification
 
 1. `pnpm dev` → `/en/`, `/es/`, `/pt/` prefixed routes render correct language
-2. `/services/web-development` → middleware redirects to `/en/services/web-development`
+2. `/services/web-development` → middleware redirects to
+   `/en/services/web-development`
 3. `/api/contact` → no redirect (stays as-is)
 4. Invalid locale `/xx/...` → 404
 5. `pnpm exec tsc --noEmit` → exit 0 (zero errors)
