+++ src/app/[locale]/page.tsx
@@ -1,11 +1,6 @@
-import Contact from "@/components/sections/Contact";
-import CTABand from "@/components/sections/CTABand";
-import FAQ from "@/components/sections/FAQ";
-import Hero from "@/components/sections/Hero";
-import Industries from "@/components/sections/Industries";
-import Services from "@/components/sections/Services";
-import StatsBar from "@/components/sections/StatsBar";
-import WhyBobatech from "@/components/sections/WhyBobatech";
+import { CTABand } from "@/shared/components";
+import { ContactForm } from "@/features/leads";
+import { Hero, StatsBar, Services, WhyBobatech, Industries, FAQ } from "@/features/home";
 import Footer from "@/components/ui/Footer";
 import Navbar from "@/components/ui/Navbar";
 import { generateOrganizationSchema } from "~/lib/seo";
@@ -29,7 +24,7 @@ export default function Home() {
 					<Industries />
 					<FAQ />
 					<CTABand />
-					<Contact />
+					<ContactForm />
 				</main>
 				<Footer />
 			</div>
