--- src/app/[locale]/pricing/constants.ts
@@ -1,1 +1,0 @@
-export { PRICING_STEPS, TIMELINE_STEP_ID, RUSH_MULTIPLIER, FLEXIBLE_MULTIPLIER, STANDARD_MULTIPLIER, ANIMATION_CONFIG } from "@/features/pricing/model/constants";
