--- src/app/[locale]/pricing/types.ts
@@ -1,1 +1,0 @@
-export type { ProjectRequirement, Step, SelectedOption, StepBreakdown, Selections } from "@/features/pricing/model/types";
