--- src/app/[locale]/pricing/validation.ts
@@ -1,1 +1,0 @@
-export { validEmail } from "@/features/pricing/model/validation";
