+++ src/app/[locale]/services/all/[industry]/page.tsx
@@ -4,7 +4,7 @@ import { notFound } from "next/navigation";
 import Navbar from "@/components/ui/Navbar";
 import Footer from "@/components/ui/Footer";
 import Button from "@/components/ui/Button";
-import { industryServices } from "@/data/services";
+import { industryServices } from "@/features/services/model/services";
 import { ArrowRight } from "lucide-react";
 import { getTranslations } from "next-intl/server";
 import { CAL_LINKS } from "~/lib/constants";
