+++ src/app/admin/_components/SignOutButton.tsx
@@ -1,14 +1,1 @@
-"use client";
-
-import { signOut } from "~/lib/auth-client";
-
-export function SignOutButton() {
-	return (
-		<button
-			onClick={() => signOut({ fetchOptions: { onSuccess: () => { window.location.href = "/admin/sign-in"; } } })}
-			className="text-xs text-brand-primary/40 hover:text-brand-primary/70 transition-colors"
-		>
-			Sign out
-		</button>
-	);
-}
+export { SignOutButton } from "@/features/admin";
