+++ src/app/sitemap.ts
@@ -1,4 +1,4 @@
-import { allServices, industryServices } from "~/data/services";
+import { allServices, industryServices } from "@/features/services/model/services";
 import { getAllPosts } from "~/lib/sanity/queries";
 
 import type { MetadataRoute } from "next";
