--- src/data/service-pages.en.ts
@@ -1,1 +1,0 @@
-export { servicePages } from "@/features/services/api/service-pages.en";
