--- src/data/service-pages.es.ts
@@ -1,1 +1,0 @@
-export { servicePages } from "@/features/services/api/service-pages.es";
