--- src/data/service-pages.pt.ts
@@ -1,1 +1,0 @@
-export { servicePages } from "@/features/services/api/service-pages.pt";
