--- src/data/service-pages.shared.ts
@@ -1,10 +1,0 @@
-export type {
-	PainPoint,
-	ServiceOffer,
-	ProcessStep,
-	TechCategory,
-	ServicePageData,
-	MvpSolution,
-	MvpForWho,
-} from "@/features/services/model/types";
-export { sharedTechStack } from "@/features/services/model/types";
