--- src/data/service-pages.ts
@@ -1,11 +1,0 @@
-export type {
-	PainPoint,
-	ServiceOffer,
-	ProcessStep,
-	TechCategory,
-	ServicePageData,
-	MvpSolution,
-	MvpForWho,
-} from "@/features/services/model/types";
-export { sharedTechStack } from "@/features/services/model/types";
-export { getServicePageData } from "@/features/services/api/getServicePage";
