--- src/data/services.ts
@@ -1,2 +1,0 @@
-export type { Service, IndustryService } from "@/features/services/model/types";
-export { allServices, industryServices } from "@/features/services/model/services";
