+++ src/features/services/components/RichServicePage.tsx
@@ -1,7 +1,7 @@
 import Navbar from "@/components/ui/Navbar";
 import Footer from "@/components/ui/Footer";
 import Button from "@/components/ui/Button";
-import CTABand from "@/components/sections/CTABand";
+import CTABand from "@/shared/components/CTABand";
 import { CAL_LINKS } from "~/lib/constants";
 import type { ServicePageData } from "@/features/services/model/types";
 import ServiceHero from "./ServiceHero";
