+++ src/features/tools/index.ts
@@ -1,0 +1,1 @@
+export { default as ToolsCatalog } from "./components/ToolsCatalog";
